backend/requirements.txt
fastapi==0.133.1
uvicorn[standard]==0.41.0
motor==3.7.1
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.1
redis==7.2.1
rq==2.7.0
miniopy-async==1.23.4
sse-starlette==3.2.0
aioredis==2.0.1
